+++ code/src/app/[lang]/registry/task/outstanding/page.tsx
@@ -1,30 +1,28 @@
-import { Metadata } from "next";
-import { redirect } from "next/navigation";
+import { Metadata } from 'next';
+import { redirect } from 'next/navigation';
 
-import { Modules, PageTitles, Paths } from "io/config/routes";
-import SettingsStore from "io/config/settings";
-import { NavBarItemSettings, UISettings } from "types/settings";
-import RegistryTableComponent from "ui/graphic/table/registry/registry-table-component";
+import { Modules, PageTitles, Paths } from 'io/config/routes';
+import SettingsStore from 'io/config/settings';
+import { NavBarItemSettings, UISettings } from 'types/settings';
+import RegistryTableComponent from 'ui/graphic/table/registry/registry-table-component';
 
 /**
  * Set page metadata.
- *
+ * 
  * @returns metadata promise.
  */
 export async function generateMetadata(): Promise<Metadata> {
   const uiSettings: UISettings = SettingsStore.getUISettings();
-  const metadata: NavBarItemSettings = uiSettings.links?.find(
-    (link) => link.url === Modules.REGISTRY
-  );
+  const metadata: NavBarItemSettings = uiSettings.links?.find(link => link.url === Modules.REGISTRY);
   return {
     title: metadata?.title ?? PageTitles.REGISTRY,
-  };
+  }
 }
 
 /**
  * Displays the registry page for viewing tasks by dates.
- *
- * @returns React component for display.
+ * 
+ * @returns React component for display. 
  */
 export default function RegistryTaskByDatePage() {
   const uiSettings: UISettings = SettingsStore.getUISettings();
@@ -32,7 +30,7 @@ export default function RegistryTaskByDatePage() {
     return (
       <RegistryTableComponent
         entityType={uiSettings.resources?.registry?.data}
-        lifecycleStage={"tasks"}
+        lifecycleStage={'outstanding'}
       />
     );
   } else {