+++ code/src/types/backend-agent.ts
@@ -7,9 +7,25 @@ export type AgentResponseBody = {
     items?: string[] | Record<string, unknown>[];
   };
   error?: {
-    code: number
+    code: number;
     message: string;
   };
 };
 
-export type InternalApiIdentifier = "address" | "concept" | "contracts" | "contract_status" | "event" | "form" | "geocode_postal" | "geocode_address" | "geocode_city" | "geodecode" | "instances" | "schedule" | "tasks"
+export type InternalApiIdentifier =
+  | "address"
+  | "concept"
+  | "contracts"
+  | "contract_status"
+  | "event"
+  | "form"
+  | "geocode_postal"
+  | "geocode_address"
+  | "geocode_city"
+  | "geodecode"
+  | "instances"
+  | "schedule"
+  | "tasks"
+  | "outstanding"
+  | "scheduled"
+  | "closed";