+++ code/src/ui/graphic/table/registry/ribbon/table-ribbon.tsx
@@ -1,10 +1,9 @@
 "use client";
 
-import React from "react";
-
 import { usePermissionScheme } from "hooks/auth/usePermissionScheme";
 import { useDictionary } from "hooks/useDictionary";
 import { Routes } from "io/config/routes";
+import React from "react";
 import { PermissionScheme } from "types/auth";
 import { Dictionary } from "types/dictionary";
 import { LifecycleStage, RegistryFieldValues } from "types/form";
@@ -12,15 +11,19 @@ import { DownloadButton } from "ui/interaction/action/download/download";
 import RedirectButton from "ui/interaction/action/redirect/redirect-button";
 import ReturnButton from "ui/interaction/action/redirect/return-button";
 import Button from "ui/interaction/button";
+import DateRangeInput from "ui/interaction/input/date-range";
 import ColumnSearchComponent from "../actions/column-search";
+import { DateRange } from "react-day-picker";
 
 interface TableRibbonProps {
   path: string;
   entityType: string;
-  selectedDate: string;
+  selectedDate: DateRange;
   lifecycleStage: LifecycleStage;
   instances: RegistryFieldValues[];
-  setSelectedDate: React.Dispatch<React.SetStateAction<string>>;
+  setSelectedDate: React.Dispatch<
+    React.SetStateAction<DateRange>
+  >;
   setCurrentInstances: React.Dispatch<
     React.SetStateAction<RegistryFieldValues[]>
   >;
@@ -32,10 +35,10 @@ interface TableRibbonProps {
  *
  * @param {string} path The current path name after the last /.
  * @param {string} entityType The type of entity.
- * @param {string} selectedDate The selected date in the date field input.
+ * @param {DateRange} selectedDate The selected date range object with 'from' and 'to' date strings.
  * @param {LifecycleStage} lifecycleStage The current stage of a contract lifecycle to display.
  * @param {RegistryFieldValues[]} instances The target instances to export into csv.
- * @param setSelectedDate Method to update selected date.
+ * @param setSelectedDate A dispatch method to update selected date range.
  * @param setCurrentInstances A dispatch method to set the current instances after parsing the initial instances.
  * @param triggerRefresh Method to trigger refresh.
  */
@@ -43,19 +46,42 @@ export default function TableRibbon(props: Readonly<TableRibbonProps>) {
   const dict: Dictionary = useDictionary();
   const keycloakEnabled = process.env.KEYCLOAK === "true";
   const permissionScheme: PermissionScheme = usePermissionScheme();
-  const taskId: string = "task date";
-
-  // Handle change event for the date input
-  const handleDateChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    props.setSelectedDate(event.target.value);
-  };
-
   const triggerRefresh: React.MouseEventHandler<HTMLButtonElement> = () => {
     props.triggerRefresh();
   };
 
   return (
     <div className="flex flex-col p-1 md:p-2 gap-2 md:gap-4">
+      {props.lifecycleStage !== "general" &&
+        props.lifecycleStage !== "pending" &&
+        (!keycloakEnabled ||
+          !permissionScheme ||
+          permissionScheme.hasPermissions.allTasks) && (
+          <div className="flex  items-centre justify-between  sm:gap-4 bg-gray-200 dark:bg-zinc-800   max-w-fit p-1.5 text-center rounded-lg flex-wrap">
+            <RedirectButton
+              label={dict.nav.title.outstanding}
+              leftIcon="pending"
+              url={`${Routes.REGISTRY_TASK_OUTSTANDING}`}
+              variant={
+                props.lifecycleStage == "outstanding" ? "active" : "ghost"
+              }
+            />
+            <RedirectButton
+              label={dict.nav.title.scheduled}
+              leftIcon="schedule"
+              url={`${Routes.REGISTRY_TASK_SCHEDULED}`}
+              variant={
+                props.lifecycleStage == "scheduled" ? "active" : "ghost"
+              }
+            />
+            <RedirectButton
+              label={dict.nav.title.closed}
+              leftIcon="archive"
+              url={`${Routes.REGISTRY_TASK_CLOSED}`}
+              variant={props.lifecycleStage == "closed" ? "active" : "ghost"}
+            />
+          </div>
+        )}
       <div className="w-full border-[0.5px] border-border" />
       <div className="flex justify-between md:gap-2 lg:gap-0 flex-wrap ">
         <div className="flex items-center !-ml-2 ">
@@ -110,29 +136,19 @@ export default function TableRibbon(props: Readonly<TableRibbonProps>) {
           {(!keycloakEnabled ||
             !permissionScheme ||
             permissionScheme.hasPermissions.export) && (
-            <DownloadButton instances={props.instances} />
-          )}
+              <DownloadButton instances={props.instances} />
+            )}
         </div>
       </div>
-      <div className="flex ml-2 ">
-        {props.lifecycleStage == "tasks" && (
-          <div className="flex items-center gap-4">
-            <label
-              className="my-1 text-sm md:text-lg text-left whitespace-nowrap"
-              htmlFor={taskId}
-            >
-              {dict.action.date}:
-            </label>
-            <input
-              id={taskId}
-              className="h-8 w-full max-w-none p-5 rounded-lg border-1 border-border bg-muted text-foreground shadow-md"
-              type={"date"}
-              defaultValue={props.selectedDate}
-              aria-label={taskId}
-              onChange={handleDateChange}
+      <div className="flex ml-2">
+        {(props.lifecycleStage == "scheduled" ||
+          props.lifecycleStage == "closed") && (
+            <DateRangeInput
+              selectedDate={props.selectedDate}
+              setSelectedDate={props.setSelectedDate}
+              lifecycleStage={props.lifecycleStage}
             />
-          </div>
-        )}
+          )}
       </div>
     </div>
   );