+++ code/src/ui/navigation/navbar/nav-menu.tsx
@@ -116,9 +116,10 @@ function NavMenuContents(
   const registryLinkProps: NavBarItemSettings = props.settings.links?.find(
     (link) => link.url === Modules.REGISTRY
   );
-  const registrySubmissionLinkProps: NavBarItemSettings = props.settings.links?.find(
-    (link) => link.url === `${Modules.REGISTRY}-submission`
-  );
+  const registrySubmissionLinkProps: NavBarItemSettings =
+    props.settings.links?.find(
+      (link) => link.url === `${Modules.REGISTRY}-submission`
+    );
 
   function createHandleFileUploadClick(
     url: string
@@ -133,20 +134,23 @@ function NavMenuContents(
 
   return (
     <div
-      className={`${props.isMobile
-        ? "flex gap-4 p-2 "
-        : "bg-muted border-r-border hidden  items-center gap-6 overflow-x-scroll overflow-y-auto border-r pb-20"
-        }
+      className={`${
+        props.isMobile
+          ? "flex gap-4 p-2 "
+          : "bg-muted border-r-border hidden  items-center gap-6 overflow-x-scroll overflow-y-auto border-r pb-20"
+      }
       ${isMenuExpanded ? "w-3xs lg:w-xs xl:flex 2xl:w-xs" : "w-24  xl:flex"}
       
          flex-col justify-start transition-all duration-200 ease-in-out`}
     >
       <button
-        className={`${props.isMobile ? "hidden" : "xl:flex"
-          }   cursor-pointer mt-4  p-4  transition-colors duration-200 hover:bg-gray-300 dark:hover:bg-zinc-700 ${isMenuExpanded
+        className={`${
+          props.isMobile ? "hidden" : "xl:flex"
+        }   cursor-pointer mt-4  p-4  transition-colors duration-200 hover:bg-gray-300 dark:hover:bg-zinc-700 ${
+          isMenuExpanded
             ? "mr-2 self-end rounded-md -mb-8 "
             : " justify-center items-center rounded-full -mb-4"
-          }`}
+        }`}
         onClick={() => setIsMenuExpanded(!isMenuExpanded)}
       >
         <Icon className="material-symbols-outlined">
@@ -228,15 +232,18 @@ function NavMenuContents(
       {props.settings.modules.registry &&
         props.settings.resources?.registry?.data &&
         (!keycloakEnabled ||
-          permissionScheme?.hasPermissions.pendingRegistry) && (
+          permissionScheme?.hasPermissions.registrySubmission) && (
           <NavBarItem
-            title={registrySubmissionLinkProps?.title ?? dict.nav.title.submission}
+            title={
+              registrySubmissionLinkProps?.title ?? dict.nav.title.submission
+            }
             icon={registrySubmissionLinkProps?.icon ?? "work"}
             url={`${Routes.REGISTRY_PENDING}/${props.settings.resources?.registry?.data}`}
             isMobile={props.isMobile}
             caption={
               isMenuExpanded
-                ? registrySubmissionLinkProps?.caption ?? dict.nav.caption.submission
+                ? registrySubmissionLinkProps?.caption ??
+                  dict.nav.caption.submission
                 : undefined
             }
             setIsOpen={props.setIsMenuOpen}
@@ -246,11 +253,12 @@ function NavMenuContents(
 
       {props.settings.modules.registry &&
         props.settings.resources?.registry?.data &&
-        (!keycloakEnabled || permissionScheme?.hasPermissions.activeArchiveRegistry) && (
+        (!keycloakEnabled ||
+          permissionScheme?.hasPermissions.registry) && (
           <NavBarItem
             title={registryLinkProps?.title ?? dict.nav.title.registry}
             icon={registryLinkProps?.icon ?? "table_chart"}
-            url={Routes.REGISTRY_TASK_DATE}
+            url={Routes.REGISTRY_TASK_OUTSTANDING}
             isMobile={props.isMobile}
             caption={
               isMenuExpanded
@@ -273,9 +281,9 @@ function NavMenuContents(
             caption={
               isMenuExpanded
                 ? dict.nav.caption.generalReg.replace(
-                  "{replace}",
-                  parseWordsForLabels(path.type).toLowerCase()
-                )
+                    "{replace}",
+                    parseWordsForLabels(path.type).toLowerCase()
+                  )
                 : undefined
             }
             setIsOpen={props.setIsMenuOpen}